refactor: Use recoil to persist chat history
